Parse log file for working directory
The working directory was assumed to be the folder
of the source file, however, this is not always be
case. Projects that search headers with
$(PROJECT_DIR) placeholder might contain relative
search path from a different directory. In this 
case, we parse build log file for 'cd' commands
and figure out the current working directory.

Thank @PROSPricing for reporting this issue.